:boom: Seed data for each entity is created

+++ StudentCredits.Domain.Entities/Details/TeacherDetailEntity.cs
@@ -5,7 +5,7 @@ public class TeacherDetailEntity
   public Guid TeacherDetailId { get; set; }
   public required Guid TeacherId { get; set; }
   public required Guid SubjectId { get; set; }
-  public required short Credits { get; set; }
+  public required decimal Credits { get; set; }
   public DateTimeOffset Created { get; set; }
   public uint Version { get; set; }
   public TeacherEntity Teacher { get; set; } = null!;
